# [ez][CH][drci] Improve recent_pr_workflows_query perf