This is the main query used by dr ci to get info on jobs.  It some times OOMs when querying:
```
(while reading column head): (while reading from part ... in table default.pull_request...
```

Remove final for push because it is not needed.
Remove final for pull_request in favor of argmax.

This brings down the time and the memory needed


```
 ~/zzzzzzzz/test-infra/tools  @c6e2eb46 *129 !2  python torchci/clickhouse_query_perf.py --query recent_pr_workflows_query  --results --perf  --base origin/main --strict                                                                                 ✔  1m 57s  forpytorch Py  09:23:07 AM 
Using unstaged changes for --head
Gathering perf stats for: recent_pr_workflows_query
Num tests: 1
Num times: 10
100%|██████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 10/10 [00:35<00:00,  3.57s/it]
100%|██████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 10/10 [00:42<00:00,  4.28s/it]
Waiting for metrics to populate, please be patient
+------+----------+-----------+-------------+---------------+-------------+-----------+--------------+--------------+
| Test | Avg Time | Base Time | Time Change | % Time Change |   Avg Mem   |  Base Mem |  Mem Change  | % Mem Change |
+------+----------+-----------+-------------+---------------+-------------+-----------+--------------+--------------+
|  0   |  2946.5  |    3769   |    -822.5   |      -22      | 135942738.5 | 565914340 | -429971601.5 |     -76      |
+------+----------+-----------+-------------+---------------+-------------+-----------+--------------+--------------+
Comparing results for query: recent_pr_workflows_query
Num tests: 1
Head: unstaged
 Base: origin/main
Results (4206 rows) for test 0 match
```